+++ CHANGELOG.md
@@ -28,6 +28,8 @@
   * machine state machine has a new state `BomValidation/SkuMissing` to handle the situation when the sku specified in the expected machine does not exist.
   * new metrics are emitted for the number of machines associated with a sku and/or device_type
 
+- [MR-4395](https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/4395) [MR-4342](https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/4342) Added ability to do host upgrades via script instead of the normal firmware update process.
+
 - [FORGE-6728](https://jirasw.nvidia.com/browse/FORGE-6728): new cli command to allow bulk update of sku device type
 
 ### Changed